Add client side api for file system

+++ server/Services/gRPC/FileSystemService.cs
@@ -35,17 +35,24 @@ namespace server.Services.gRPC
         {
             authHelper.EnsurePermissionForComputer(context, request.Computer.ToGuid());
 
+            if (File.IsNameValid(request.FileName) == false)
+            {
+                throw new RpcException(new Status(StatusCode.InvalidArgument, "Invalid file name"));
+            }
+
+            transCtrl.BeginTrans();
             Guid computer = request.Computer.ToGuid();
             Guid parent = request.Parent.ToGuid();
             if (fileSystemManager.ExistsDirectory(computer, parent) == false)
             {
+                transCtrl.Rollback();
                 throw new RpcException(new Status(StatusCode.NotFound, "Parent directory not found"));
             }
             if (fileSystemManager.ExistsFile(computer, parent, request.FileName))
             {
+                transCtrl.Rollback();
                 throw new RpcException(new Status(StatusCode.AlreadyExists, "File already exists"));
             }
-            transCtrl.BeginTrans();
             var id = fileSystemManager.CreateFile(computer, parent, request.FileName, FileType.Readable | FileType.Writable);
             fileDataManager.Create(computer, id, request.Data.ToByteArray());
             transCtrl.Commit();
@@ -56,17 +63,26 @@ namespace server.Services.gRPC
         {
             authHelper.EnsurePermissionForComputer(context, request.Computer.ToGuid());
 
+            if (Directory.IsNameValid(request.DirectoryName) == false)
+            {
+                throw new RpcException(new Status(StatusCode.InvalidArgument, "Invalid directory name"));
+            }
+
+            transCtrl.BeginTrans();
             Guid computer = request.Computer.ToGuid();
             Guid parent = request.Parent.ToGuid();
             if (fileSystemManager.ExistsDirectory(computer, parent) == false)
             {
+                transCtrl.Rollback();
                 throw new RpcException(new Status(StatusCode.NotFound, "Parent directory not found"));
             }
             if (fileSystemManager.ExistsDirectory(computer, parent, request.DirectoryName))
             {
+                transCtrl.Rollback();
                 throw new RpcException(new Status(StatusCode.AlreadyExists, "Directory already exists"));
             }
             var id = fileSystemManager.CreateDirectory(computer, parent, request.DirectoryName);
+            transCtrl.Commit();
             return Task<CreateDirectoryRespond>.FromResult(new CreateDirectoryRespond { Directory = id.ToByteString() });
         }
 
@@ -74,15 +90,16 @@ namespace server.Services.gRPC
         {
             authHelper.EnsurePermissionForComputer(context, request.Computer.ToGuid());
 
+            transCtrl.BeginTrans();
             Guid computer = request.Computer.ToGuid();
             Guid id = request.Directory.ToGuid();
             if (fileSystemManager.ExistsDirectory(computer, id) == false)
             {
+                transCtrl.Rollback();
                 throw new RpcException(new Status(StatusCode.NotFound, "Directory not found"));
             }
             if (request.Recursive)
             {
-                transCtrl.BeginTrans();
                 IEnumerable<Guid> childDirectories = new List<Guid>() { id };
                 IEnumerable<Guid> childFiles = new List<Guid>();
                 while (childDirectories.Any() || childFiles.Any())
@@ -110,9 +127,11 @@ namespace server.Services.gRPC
             {
                 if (fileSystemManager.IsDirectoryEmpty(computer, id) == false)
                 {
+                    transCtrl.Rollback();
                     throw new RpcException(new Status(StatusCode.FailedPrecondition, "Directory is not empty"));
                 }
                 fileSystemManager.DeleteDirectory(computer, id);
+                transCtrl.Commit();
                 return Task<DeleteDirectoryRespond>.FromResult(new DeleteDirectoryRespond());
             }
         }
@@ -121,13 +140,19 @@ namespace server.Services.gRPC
         {
             authHelper.EnsurePermissionForComputer(context, request.Computer.ToGuid());
 
+            transCtrl.BeginTrans();
             Guid computer = request.Computer.ToGuid();
             Guid id = request.File.ToGuid();
-            var file = fileSystemManager.GetFileById(computer, id)
-            ?? throw new RpcException(new Status(StatusCode.NotFound, "File not found"));
+            var file = fileSystemManager.GetFileById(computer, id);
+            if (file == null)
+            {
+                transCtrl.Rollback();
+                throw new RpcException(new Status(StatusCode.NotFound, "File not found"));
+            }
             fileSystemManager.DeleteFile(computer, id);
             if (file.Type.HasFlag(FileType.Readable) && file.Type.HasFlag(FileType.Writable))
                 fileDataManager.Delete(computer, id);
+            transCtrl.Commit();
             return Task<DeleteFileRespond>.FromResult(new DeleteFileRespond());
         }
 
@@ -135,13 +160,13 @@ namespace server.Services.gRPC
         {
             authHelper.EnsurePermissionForComputer(context, request.Computer.ToGuid());
 
+            transCtrl.BeginTrans(); // This operation is read-only, so no need to commit or rollback
             Guid computer = request.Computer.ToGuid();
             Guid id = request.Id.ToGuid();
             if (fileSystemManager.ExistsDirectory(computer, id))
             {
                 StringBuilder stringBuilder = new();
-                Directory dir = fileSystemManager.GetDirectoryById(computer, id)
-                    ?? throw new RpcException(new Status(StatusCode.Internal, "Directory is deleted during invoke"));
+                Directory dir = fileSystemManager.GetDirectoryById(computer, id)!;
                 if (dir.Parent == Guid.Empty)
                     stringBuilder.Insert(0, "/");
                 while (dir.Parent != Guid.Empty)
@@ -156,8 +181,7 @@ namespace server.Services.gRPC
             else if (fileSystemManager.ExistsFile(computer, id))
             {
                 StringBuilder stringBuilder = new();
-                File file = fileSystemManager.GetFileById(computer, id)
-                    ?? throw new RpcException(new Status(StatusCode.Internal, "Directory is deleted during invoke"));
+                File file = fileSystemManager.GetFileById(computer, id)!;
                 stringBuilder.Insert(0, file.Name);
                 stringBuilder.Insert(0, '/');
                 Directory dir = fileSystemManager.GetDirectoryById(computer, file.Parent)!;
@@ -180,12 +204,14 @@ namespace server.Services.gRPC
         {
             authHelper.EnsurePermissionForComputer(context, request.Computer.ToGuid());
 
+            transCtrl.BeginTrans();
             Guid computer = request.Computer.ToGuid();
             Guid rootDir = computerManager.QueryComputerById(computer).RootDirectory;
             Guid startDir = request.StartDirectory.ToGuid();
             string[] path = request.Path.Split('/');
-            transCtrl.BeginTrans();
-            if (path.Length == 0) return Task<FromPathToIdRespond>.FromResult(new FromPathToIdRespond() { IsDirectory = true, Id = rootDir.ToByteString() });
+            if (fileSystemManager.ExistsDirectory(computer, request.StartDirectory.ToGuid()) == false)
+                throw new RpcException(new Status(StatusCode.NotFound, "Start directory not found"));
+            if (path.Length == 0) return Task.FromResult(new FromPathToIdRespond() { IsDirectory = true, Id = rootDir.ToByteString() });
             if (path.Length == 1)
             {
                 if (string.IsNullOrEmpty(path[0])) // case ""
@@ -267,6 +293,7 @@ namespace server.Services.gRPC
         {
             authHelper.EnsurePermissionForComputer(context, request.Computer.ToGuid());
 
+            transCtrl.BeginTrans(); // This is a read-only operation
             Guid computer = request.Computer.ToGuid();
             Guid directory = request.Directory.ToGuid();
             var dir = fileSystemManager.GetDirectoryById(computer, directory)
@@ -278,6 +305,7 @@ namespace server.Services.gRPC
         {
             authHelper.EnsurePermissionForComputer(context, request.Computer.ToGuid());
 
+            transCtrl.BeginTrans(); // This is a read-only operation
             Guid computer = request.Computer.ToGuid();
             Guid file = request.File.ToGuid();
             var f = fileSystemManager.GetFileById(computer, file)
@@ -289,8 +317,11 @@ namespace server.Services.gRPC
         {
             authHelper.EnsurePermissionForComputer(context, request.Computer.ToGuid());
 
+            transCtrl.BeginTrans();
             Guid computer = request.Computer.ToGuid();
             Guid directory = request.Directory.ToGuid();
+            if (fileSystemManager.ExistsDirectory(computer, directory) == false)
+                throw new RpcException(new Status(StatusCode.NotFound, "Directory not exists"));
             var dirs = fileSystemManager.ListDirectories(computer, directory);
             if (dirs.IsNullOrEmpty())
                 return Task<ListDirectoriesRespond>.FromResult(new ListDirectoriesRespond() { Infos = { } });
@@ -302,8 +333,11 @@ namespace server.Services.gRPC
         {
             authHelper.EnsurePermissionForComputer(context, request.Computer.ToGuid());
 
+            transCtrl.BeginTrans();
             Guid computer = request.Computer.ToGuid();
             Guid directory = request.Directory.ToGuid();
+            if (fileSystemManager.ExistsDirectory(computer, directory) == false)
+                throw new RpcException(new Status(StatusCode.NotFound, "Directory not exists"));
             var files = fileSystemManager.ListFiles(computer, directory);
             if (files.IsNullOrEmpty())
                 return Task<ListFilesRespond>.FromResult(new ListFilesRespond() { Infos = { } });
@@ -315,14 +349,23 @@ namespace server.Services.gRPC
         {
             authHelper.EnsurePermissionForComputer(context, request.Computer.ToGuid());
 
+            transCtrl.BeginTrans();
             Guid computer = request.Computer.ToGuid();
             Guid file = request.File.ToGuid();
-            var f = fileSystemManager.GetFileById(computer, file)
-                ?? throw new RpcException(new Status(StatusCode.NotFound, "File not exists"));
+            var f = fileSystemManager.GetFileById(computer, file);
+            if (f == null)
+            {
+                transCtrl.Rollback();
+                throw new RpcException(new Status(StatusCode.NotFound, "File not exists"));
+            }
             if (false == f.Type.HasFlag(File.TypeCode.Writable)
                 || false == f.Type.HasFlag(File.TypeCode.Readable))
+            {
+                transCtrl.Rollback();
                 throw new RpcException(new Status(StatusCode.InvalidArgument, "File is not a data file"));
+            }
             fileDataManager.Modify(computer, file, request.Data.ToArray());
+            transCtrl.Commit();
             return Task<ModifyDataFileRespond>.FromResult(new ModifyDataFileRespond());
         }
 
@@ -330,6 +373,7 @@ namespace server.Services.gRPC
         {
             authHelper.EnsurePermissionForComputer(context, request.Computer.ToGuid());
 
+            transCtrl.BeginTrans(); // This is a read-only operation
             Guid computer = request.Computer.ToGuid();
             Guid file = request.File.ToGuid();
             var f = fileSystemManager.GetFileById(computer, file)
@@ -343,11 +387,25 @@ namespace server.Services.gRPC
         {
             authHelper.EnsurePermissionForComputer(context, request.Computer.ToGuid());
 
+            if (Directory.IsNameValid(request.Name) == false)
+                throw new RpcException(new Status(StatusCode.InvalidArgument, "Invalid directory name"));
+
+            transCtrl.BeginTrans();
             Guid computer = request.Computer.ToGuid();
             Guid directory = request.Directory.ToGuid();
-            _ = fileSystemManager.GetDirectoryById(computer, directory)
-                ?? throw new RpcException(new Status(StatusCode.NotFound, "Directory not exists"));
+            var dir = fileSystemManager.GetDirectoryById(computer, directory);
+            if (dir == null)
+            {
+                transCtrl.Rollback();
+                throw new RpcException(new Status(StatusCode.NotFound, "Directory not exists"));
+            }
+            if (fileSystemManager.GetDirectoryByName(computer, dir.Parent, request.Name) != null)
+            {
+                transCtrl.Rollback();
+                throw new RpcException(new Status(StatusCode.AlreadyExists, "Directory already exists"));
+            }
             fileSystemManager.RenameDirectory(computer, directory, request.Name);
+            transCtrl.Commit();
             return Task<RenameDirectoryRespond>.FromResult(new RenameDirectoryRespond());
         }
 
@@ -355,11 +413,25 @@ namespace server.Services.gRPC
         {
             authHelper.EnsurePermissionForComputer(context, request.Computer.ToGuid());
 
+            if (File.IsNameValid(request.Name) == false)
+                throw new RpcException(new Status(StatusCode.InvalidArgument, "Invalid file name"));
+
+            transCtrl.BeginTrans();
             Guid computer = request.Computer.ToGuid();
             Guid file = request.File.ToGuid();
-            _ = fileSystemManager.GetFileById(computer, file)
-                ?? throw new RpcException(new Status(StatusCode.NotFound, "File not exists"));
+            var f = fileSystemManager.GetFileById(computer, file);
+            if (f == null)
+            {
+                transCtrl.Rollback();
+                throw new RpcException(new Status(StatusCode.NotFound, "File not exists"));
+            }
+            if (fileSystemManager.GetFileByName(computer, f.Parent, request.Name) != null)
+            {
+                transCtrl.Rollback();
+                throw new RpcException(new Status(StatusCode.AlreadyExists, "File already exists"));
+            }
             fileSystemManager.RenameFile(computer, file, request.Name);
+            transCtrl.Commit();
             return Task<RenameFileRespond>.FromResult(new RenameFileRespond());
         }
     }
